+++ src/AbstractIntType.php
@@ -6,18 +6,22 @@ use Fortuneglobe\Types\Exceptions\InvalidIntValueException;
 use Fortuneglobe\Types\Interfaces\RepresentsIntValue;
 
 /**
- * Class IntType
+ * Class AbstractIntType
  * @package Fortuneglobe\Types
  */
-abstract class IntType extends AbstractType implements RepresentsIntValue
+abstract class AbstractIntType extends AbstractType implements RepresentsIntValue
 {
 	/** @var int */
 	private $value;
 
 	public function __construct( int $value )
 	{
+		$this->guardValueIsValid( $value );
+
 		$this->value = $value;
 	}
+
+	abstract protected function guardValueIsValid( int $value ) : void;
 
 	public function toString() : string
 	{