+++ tests/Unit/Fixtures/TestFloat.php
@@ -5,13 +5,16 @@
 
 namespace Fortuneglobe\Types\Tests\Unit\Fixtures;
 
-use Fortuneglobe\Types\FloatType;
+use Fortuneglobe\Types\AbstractFloatType;
 
 /**
  * Class TestFloat
  * @package Fortuneglobe\Types\Tests\Unit\Fixtures
  */
-final class TestFloat extends FloatType
+final class TestFloat extends AbstractFloatType
 {
-
+	protected function guardValueIsValid( float $value ) : void
+	{
+		// TODO: Implement guardValueIsValid() method.
+	}
 }