+++ CHANGELOG.md
@@ -1,8 +1,25 @@
 # CHANGELOG
 
-## [1.0.0] - YYYY-MM-DD
+## [1.0.0] - 2017-11-13
+
+### Changed
+
+* All base types prefixed with `Abstract`
+* Renamed `UuidType` to `AbstractUuid4Type` to clarify UUID version
+* All base types define an abstract `guardValueIsValid()` method. (Except `AbstractUuid4Type`)
+
+## [0.9.1] - 2017-10-30
+
+### Fixed
+
+* Composer package name and description - [#2]
+
+## [0.9.0] - 2017-10-26
 
 * Initial release
 
-[1.0.0]: https://github.com/fortuneglobe/types/tags/v1.0.0
+[1.0.0]: https://github.com/fortuneglobe/types/compare/v0.9.1...v1.0.0
+[0.9.1]: https://github.com/fortuneglobe/types/compare/v0.9.0...v0.9.1
+[0.9.0]: https://github.com/fortuneglobe/types/tags/v0.9.0
 
+[#2]: https://github.com/fortuneglobe/types/issues/2